refactor: Extract consent and callback logic into oidc_consent.go
Move ConsentEntry, ConsentCodeConsumer, ConsentInfoResponse types,
HandleConsentInfo, handleConsentCallback, handleDexCallback,
buildConsentRedirect, and buildConsentPageURL into oidc_consent.go
to keep oidc.go under the 800-line service convention limit.
